@@ -18,44 +18,45 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^1.0.0
-  styled_widget: ^0.2.2
-  filepicker_windows: ^1.0.0
-  xml: ^4.5.1
-  get_it: ^5.0.1
-  flutter_spinkit: ^4.1.2+1
-  flutter_typeahead: ^1.8.8
-  json_annotation: ^3.1.0
-  freezed_annotation: ^0.12.0
-  flutter_simple_treeview: ^2.0.2
-  recase: ^3.0.0
-  mobx: ^1.2.1+4
-  flutter_mobx: ^1.1.0+2
-  shared_preferences: ^0.5.12+2
-  provider: ^4.3.2+2
-  flutter_form_builder: ^3.14.1
-  encrypt: ^4.1.0
-  path_provider: ^1.6.22
-  uuid: ^2.2.2
-  graphview: ^0.6.1
-  split_view: ^1.0.1
-  flutter_highlight: ^0.6.0
+  styled_widget: 
+    git: 
+      url: git://github.com/vance-liu/styled_widget.git
+      ref: null_safety
+  collection: ^1.15.0
+  filepicker_windows: ^2.0.0-nullsafety.0
+  xml: ^5.0.2
+  get_it: ^6.0.0
+  flutter_spinkit: ^5.0.0
+  json_annotation: ^4.0.1
+  freezed_annotation: ^0.14.1
+  flutter_simple_treeview: ^3.0.0-nullsafety.1
+  recase: ^4.0.0-nullsafety.0
+  mobx: ^2.0.1
+  flutter_mobx: ^2.0.0
+  shared_preferences: ^2.0.5
+  provider: ^5.0.0
+  flutter_form_builder: ^6.0.0-nullsafety.1
+  encrypt: ^5.0.0
+  path_provider: ^2.0.1
+  uuid: ^3.0.3
+  flutter_highlight: ^0.7.0
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   
-  build_runner: ^1.10.4
-  json_serializable: ^3.5.0
-  freezed: ^0.12.2
-  mobx_codegen: ^1.1.2
+  build_runner: ^1.12.2
+  json_serializable: ^4.1.0
+  freezed: ^0.14.1+2
+  mobx_codegen: ^2.0.1+2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
