@@ -1,31 +1,32 @@
 import 'package:flutter/material.dart';
 
-Future<T> confirm<T>(
+Future<bool> confirm(
   BuildContext context, {
-  Widget title,
-  Widget content,
-  Widget textOK,
-  Widget textCancel,
-}) {
-  return showDialog(
+  Widget? title,
+  required Widget content,
+  Widget textOK = const Text('Si'),
+  Widget textCancel = const Text('No'),
+}) async {
+  final result = await showDialog<bool>(
     context: context,
     builder: (_) => WillPopScope(
       child: AlertDialog(
-        title: title ?? Text(''),
+        title: title,
         content: content,
         actions: <Widget>[
-          FlatButton(
-              child: textCancel ?? Text('No'),
+          TextButton(
+              child: textCancel,
               onPressed: () => Navigator.pop(context, false)),
-          FlatButton(
-              child: textOK ?? Text('Si'),
-              onPressed: () => Navigator.pop(context, true)),
+          TextButton(
+              child: textOK, onPressed: () => Navigator.pop(context, true)),
         ],
       ),
-      onWillPop: () {
+      onWillPop: () async {
         Navigator.pop(context, false);
-        return;
+        return false;
       },
     ),
   );
+
+  return result ?? false;
 }
