@@ -2,15 +2,15 @@ import 'package:flutter/material.dart';
 import 'package:styled_widget/styled_widget.dart';
 
 class ActionButton extends StatelessWidget {
-  ActionButton({this.icon, this.label, this.onPressed});
+  ActionButton({required this.icon, required this.label, this.onPressed});
 
   final IconData icon;
   final String label;
-  final Function onPressed;
+  final void Function()? onPressed;
 
   @override
   Widget build(BuildContext context) {
-    return FlatButton.icon(
+    return TextButton.icon(
       icon: Icon(icon).iconColor(Colors.white).semanticsLabel(label),
       label: Text(label).textColor(Colors.white),
       onPressed: onPressed,
