@@ -6,25 +6,25 @@ import '../common/extensions.dart';
 class Indicator extends StatelessWidget {
   Indicator(
     this.content, {
-    Key key,
+    Key? key,
     this.icon,
-    this.size,
+    this.size = 24.0,
     this.color,
-    this.alignment,
+    this.alignment = MainAxisAlignment.start,
   }) : super(key: key);
 
   final Text content;
-  final IconData icon;
+  final IconData? icon;
   final double size;
-  final Color color;
+  final Color? color;
   final MainAxisAlignment alignment;
 
   @override
   Widget build(BuildContext context) {
     final color = this.color ?? Theme.of(context).primaryIconTheme.color;
-    final size = this.size ?? 24.0;
+    final size = this.size;
     return Row(
-      mainAxisAlignment: alignment ?? MainAxisAlignment.start,
+      mainAxisAlignment: alignment,
       children: [
         icon != null
             ? Icon(icon, color: color, size: size)
