@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:styled_widget/styled_widget.dart';
+import 'package:collection/collection.dart';
 
 import 'package:migrator/common/common.dart';
 import 'package:migrator/models/models.dart';
@@ -8,7 +9,7 @@ import 'package:migrator/stores/stores.dart';
 
 class MigrateScreenBody extends StatefulWidget {
   MigrateScreenBody({
-    this.statusBar,
+    required this.statusBar,
     this.mappingActionEditable = false,
     this.headersHook,
     this.rowsHook,
@@ -16,14 +17,14 @@ class MigrateScreenBody extends StatefulWidget {
   });
 
   final Widget statusBar;
-  final Function(List<String> labels) headersHook;
-  final Function(List<Widget> cells, ItemWithId item) rowsHook;
+  final Function(List<String> labels)? headersHook;
+  final Function(List<Widget> cells, ItemWithId? item)? rowsHook;
   final bool mappingActionEditable;
   final Widget Function(
     ClusterPropertyItem cwp,
     String value,
     bool isOverflow,
-  ) cwpSuffixIconBuilder;
+  )? cwpSuffixIconBuilder;
 
   @override
   _MigrateScreenBodyState createState() => _MigrateScreenBodyState();
@@ -113,9 +114,9 @@ class _MigrateScreenBodyState extends State<MigrateScreenBody> {
 
   Widget _buildTable(
     BuildContext context, {
-    @required List<ItemMapping> mappings,
-    ItemType only,
-    ItemType ignore,
+    required List<ItemMapping> mappings,
+    ItemType? only,
+    ItemType? ignore,
   }) {
     final isForCwp = only == ItemType.clusterProperty;
     final header = _buildTableHeader(isForCwp);
@@ -149,7 +150,7 @@ class _MigrateScreenBodyState extends State<MigrateScreenBody> {
   TableRow _buildTableHeader(bool isForCwp) {
     final labels = ['Nombre', (isForCwp ? 'Valor' : 'Tipo'), 'Acción'];
     if (widget.headersHook != null) {
-      widget.headersHook(labels);
+      widget.headersHook!(labels);
     }
     final headers = labels
         .map((text) => Text(text).fontSize(16.0).fontWeight(FontWeight.bold));
@@ -159,40 +160,40 @@ class _MigrateScreenBodyState extends State<MigrateScreenBody> {
 
   TableRow _buildTableRow(BuildContext context, ItemMapping mapping) {
     final store = context.store<MigrateStore>();
-    var item = store.bundle.items
-        .firstWhere((item) => item.id == mapping.srcId, orElse: () => null);
+    var item = store.bundle?.items
+        .firstWhereOrNull((item) => item.id == mapping.srcId);
 
     if (item == null && mapping.type == ItemType.folder) {
-      item = store.folders.firstWhere(
-        (folder) => folder.id == mapping.srcId,
-        orElse: () => null,
-      );
+      item = store.folders
+          .firstWhereOrNull((folder) => folder.id == mapping.srcId);
     }
+
+    List<Widget> cells;
 
     if (item == null) {
       print(
         'Item is not present in the bundle [id: ${mapping.srcId} type: ${mapping.rawType}]',
       );
-    }
 
-    final cells = [
-      _buildItemNameCell(item),
-      mapping.type == ItemType.clusterProperty
-          ? _buildClusterPropertyCell(item)
-          : _buildItemTypeCell(mapping),
-      _buildActionMappingCell(item)
-    ];
+      cells = [Text('---'), Text('---'), Text('---')];
+    } else {
+      cells = [
+        _buildItemNameCell(item),
+        item is ClusterPropertyItem
+            ? _buildClusterPropertyCell(item)
+            : _buildItemTypeCell(mapping),
+        _buildActionMappingCell(item)
+      ];
+    }
 
     if (widget.rowsHook != null) {
-      widget.rowsHook(cells, item);
+      widget.rowsHook!(cells, item);
     }
 
-    return TableRow(
-      children: [...cells],
-    );
+    return TableRow(children: [...cells]);
   }
 
-  Widget _buildItemNameCell(Item item) {
+  Widget _buildItemNameCell(ItemWithId item) {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [_buildItemNameText(item), _buildItemIdText(item)],
@@ -204,18 +205,16 @@ class _MigrateScreenBodyState extends State<MigrateScreenBody> {
   }
 
   Widget _buildItemNameText(Item item) {
-    var name = item?.name ?? 'Unknown';
+    var name = item.name;
 
-    if (item is ServiceItem) {
-      name += ' [${item.urlPattern}]';
-    }
+    if (item is ServiceItem) name += ' [${item.urlPattern}]';
 
     return Text(name);
   }
 
   Widget _buildItemIdText(ItemWithId item) {
     return Text(
-      'ID: ${item?.id ?? ''}',
+      'ID: ${item.id}',
       style: TextStyle(
         fontSize: 10.0,
         fontStyle: FontStyle.italic,
@@ -230,7 +229,7 @@ class _MigrateScreenBodyState extends State<MigrateScreenBody> {
       final value = store.clusterProperties.get(
         cwp.id,
         orElse: () => cwp.value,
-      );
+      )!;
       final maxLength = 45;
       final isOverflow = value.length > maxLength;
       return Row(
@@ -238,24 +237,29 @@ class _MigrateScreenBodyState extends State<MigrateScreenBody> {
           Text(isOverflow ? value.substring(0, maxLength) + '...' : value),
           SizedBox(width: 5.0),
           widget.cwpSuffixIconBuilder != null
-              ? widget.cwpSuffixIconBuilder(cwp, value, isOverflow)
+              ? widget.cwpSuffixIconBuilder!(cwp, value, isOverflow)
               : SizedBox()
         ],
       );
     });
   }
 
-  Widget _buildActionMappingCell(ItemWithId item) {
+  Widget _buildActionMappingCell(ItemWithId? item) {
     if (item == null) return Text('Unknown');
 
     final store = context.store<MigrateStore>();
-    final customMapping = store.mappings.get(item.id);
-    final mappingAction = customMapping?.action ?? MappingAction.ignore;
+    final customMapping = store.mappings.get(
+      item.id,
+      orElse: () => MappingConfig(action: MappingAction.unknown),
+    )!;
+    final mappingAction = customMapping.action;
     if (widget.mappingActionEditable) {
       return DropdownButton<MappingAction>(
         value: mappingAction,
         underline: SizedBox(),
-        onChanged: (action) => store.setMappingAction(item, action),
+        onChanged: (action) {
+          if (action != null) store.setMappingAction(item, action);
+        },
         items: MappingAction.values
             .map(
               (value) => DropdownMenuItem(
