@@ -5,30 +5,26 @@ final key = Key.fromUtf8(pwd);
 final iv = IV.fromLength(16);
 final encrypter = Encrypter(AES(key));
 
-String encrypt(String plainText, {throwError: true}) {
+String? encrypt(String plainText, {String orElse(Object error)?}) {
   try {
     final encrypted = encrypter.encrypt(plainText, iv: iv);
 
     return encrypted.base64;
   } catch (err) {
-    if (throwError) {
-      throw err;
-    }
-  }
+    if (orElse == null) throw err;
 
-  return null;
+    return orElse(err);
+  }
 }
 
-String decrypt(String encoded, {throwError: true}) {
+String decrypt(String encoded, {String orElse(Object error)?}) {
   try {
     final decrypted = encrypter.decrypt64(encoded, iv: iv);
 
     return decrypted;
   } catch (err) {
-    if (throwError) {
-      throw err;
-    }
-  }
+    if (orElse == null) throw err;
 
-  return null;
+    return orElse(err);
+  }
 }
