@@ -8,30 +8,24 @@ import 'item_mapping.dart';
 class BundleItem extends Item {
   BundleItem(XmlElement element) : super(element);
 
-  XmlElement get _bundle =>
-      element.getElement('l7:Resource').getElement('l7:Bundle');
+  XmlElement? get _bundle =>
+      element.getElement('l7:Resource')?.getElement('l7:Bundle');
 
   List<ItemWithId> get items {
     return _bundle
-        .getElement('l7:References')
-        .findElements('l7:Item')
-        .map((e) => ItemFactory.fromElement<ItemWithId>(e))
-        .toList();
+            ?.getElement('l7:References')
+            ?.findElements('l7:Item')
+            .map((e) => ItemFactory.fromElement<ItemWithId>(e))
+            .toList() ??
+        [];
   }
 
-  List<ItemMapping> get mappings {
-    return _bundle
-        .getElement('l7:Mappings')
-        .findElements('l7:Mapping')
-        .map((e) {
-          try {
-            return ItemMapping(e);
-          } catch (err) {
-            print('Error in: $e');
-            return null;
-          }
-        })
-        .where((x) => x != null)
-        .toList();
-  }
+  List<ItemMapping> get mappings =>
+      _bundle
+          ?.getElement('l7:Mappings')
+          ?.findElements('l7:Mapping')
+          .map((e) => ItemMapping(e))
+          .where((x) => x.srcId.isNotEmpty)
+          .toList() ??
+      [];
 }
