@@ -5,10 +5,8 @@ import 'item_with_id.dart';
 class ClusterPropertyItem extends ItemWithId {
   ClusterPropertyItem(XmlElement element) : super(element);
 
-  XmlElement get _cpElement =>
-      element.getElement('l7:Resource').getElement('l7:ClusterProperty');
+  XmlElement? get _cpElement =>
+      element.getElement('l7:Resource')?.getElement('l7:ClusterProperty');
 
-  String get value {
-    return _cpElement?.getElement('l7:Value')?.text;
-  }
+  String get value => _cpElement?.getElement('l7:Value')?.text ?? '';
 }
