@@ -1,20 +1,17 @@
 import 'item.dart';
 
 class ItemConstant {
-  const ItemConstant({this.title, this.resource});
+  const ItemConstant(this.title, this.resource);
 
   final String title;
   final String resource;
 
   static const Map<ItemType, ItemConstant> _constants = {
-    ItemType.folder: const ItemConstant(title: 'Carpetas', resource: 'folders'),
-    ItemType.service:
-        const ItemConstant(title: 'Servicios', resource: 'services'),
-    ItemType.policy:
-        const ItemConstant(title: 'Politicas', resource: 'policies'),
+    ItemType.folder: const ItemConstant('Carpetas', 'folders'),
+    ItemType.service: const ItemConstant('Servicios', 'services'),
+    ItemType.policy: const ItemConstant('Politicas', 'policies'),
   };
 
-  static ItemConstant of(ItemType type) => _constants.containsKey(type)
-      ? _constants[type]
-      : ItemConstant(title: "", resource: "");
+  static ItemConstant of(ItemType type) =>
+      _constants[type] ?? ItemConstant("", "");
 }
