@@ -16,16 +16,19 @@ class ItemFactory {
   static List<T> listFromXml<T extends Item>(String xml) {
     final doc = XmlDocument.parse(xml);
     final items = doc.firstElementChild
-        .findElements('l7:Item')
+        ?.findElements('l7:Item')
         .map((e) => fromElement<T>(e))
         .toList();
 
-    return items;
+    return items ?? [];
   }
 
   static T fromXml<T extends Item>(String xml) {
     final doc = XmlDocument.parse(xml);
     final element = doc.firstElementChild;
+
+    if (element == null)
+      throw Exception('Unexpected xml in ItemFactory.fromXml');
 
     return fromElement<T>(element);
   }
@@ -38,9 +41,11 @@ class ItemFactory {
       throw '$errorType: $errorDetail';
     }
 
-    final rawType = element.getElement('l7:Type')?.text;
-
-    final type = parseEnum(ItemType.values, rawType.toCamelCase());
+    final type = parseEnum(
+      ItemType.values,
+      (element.getElement('l7:Type')?.text ?? '').toCamelCase(),
+      orElse: () => ItemType.unknown,
+    );
 
     Item item;
 
